# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ManageHR API",
        "description": "Documentación de la API RESTful de ManageHR desarrollada en Laravel con autenticación JWT",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/experienciaLaboral": {
            "get": {
                "tags": [
                    "Experiencia Laboral"
                ],
                "summary": "Listar todas las experiencias laborales",
                "operationId": "02606ee955d45e8a3122b15846fca9f8",
                "responses": {
                    "200": {
                        "description": "Listado de experiencias laborales de los empleados "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Experiencia Laboral"
                ],
                "summary": "Crear nueva experiencia laboral",
                "operationId": "275fc232e00ebdb1b8f72d53872505fc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nomEmpresa",
                                    "nomJefe",
                                    "telefono",
                                    "cargo",
                                    "actividades",
                                    "fechaInicio",
                                    "fechaFinalizacion"
                                ],
                                "properties": {
                                    "nomEmpresa": {
                                        "type": "string",
                                        "maxLength": 45
                                    },
                                    "nomJefe": {
                                        "type": "string",
                                        "maxLength": 45
                                    },
                                    "telefono": {
                                        "type": "integer"
                                    },
                                    "cargo": {
                                        "type": "string",
                                        "maxLength": 20
                                    },
                                    "actividades": {
                                        "type": "string"
                                    },
                                    "fechaInicio": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "fechaFinalizacion": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Experiencia creada"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/experienciaLaboral/archivo": {
            "post": {
                "tags": [
                    "Experiencia Laboral"
                ],
                "summary": "Registrar experiencia laboral con archivo y relación con hoja de vida",
                "operationId": "d24846b3e478ef8591f78b40c1a740c9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nomEmpresa",
                                    "nomJefe",
                                    "telefono",
                                    "cargo",
                                    "fechaInicio",
                                    "fechaFinalizacion",
                                    "idHojaDevida"
                                ],
                                "properties": {
                                    "nomEmpresa": {
                                        "type": "string"
                                    },
                                    "nomJefe": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "number"
                                    },
                                    "cargo": {
                                        "type": "string"
                                    },
                                    "actividades": {
                                        "type": "string"
                                    },
                                    "fechaInicio": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "fechaFinalizacion": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "idHojaDevida": {
                                        "type": "integer"
                                    },
                                    "archivo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Experiencia con archivo registrada correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/experienciaLaboral/{id}": {
            "get": {
                "tags": [
                    "Experiencia Laboral"
                ],
                "summary": "Obtener una experiencia laboral por ID",
                "operationId": "0868f668bc856bfdba657a3f3e761ad8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la experiencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Experiencia encontrada"
                    },
                    "404": {
                        "description": "Experiencia no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Experiencia Laboral"
                ],
                "summary": "Actualizar una experiencia laboral",
                "operationId": "e45e0c6e1f0462777451bcf00c57e137",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la experiencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nomEmpresa",
                                    "nombJefe",
                                    "telefono",
                                    "cargo",
                                    "actividades",
                                    "fechaInicio",
                                    "fechaFinalizacion"
                                ],
                                "properties": {
                                    "nomEmpresa": {
                                        "type": "string"
                                    },
                                    "nombJefe": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "number"
                                    },
                                    "cargo": {
                                        "type": "string"
                                    },
                                    "actividades": {
                                        "type": "string"
                                    },
                                    "fechaInicio": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "fechaFinalizacion": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Experiencia actualizada correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Experiencia Laboral"
                ],
                "summary": "Eliminar una experiencia laboral",
                "operationId": "b566b727c5325d032dde81cd3f4aba20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la experiencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Experiencia eliminada correctamente"
                    },
                    "404": {
                        "description": "Experiencia no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Experiencia Laboral"
                ],
                "summary": "Actualizar parcialmente una experiencia laboral",
                "operationId": "c4e5bfea141f3ebd00f41be7771f8078",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la experiencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nomEmpresa": {
                                        "type": "string"
                                    },
                                    "nombJefe": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "number"
                                    },
                                    "cargo": {
                                        "type": "string"
                                    },
                                    "actividades": {
                                        "type": "string"
                                    },
                                    "fechaInicio": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "fechaFinalizacion": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Experiencia actualizada parcialmente"
                    },
                    "400": {
                        "description": "Error en la validación parcial"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Introduce tu token JWT en el formato: Bearer {token}"
            }
        }
    },
...
}